+++ src/Pages/Upgrades/Components/index.js
@@ -1,0 +1,1 @@
+export { default as UpgradesList } from './UpgradesList';