+++ src/redux/actions/index.js
@@ -10,7 +10,7 @@ import * as validatorsActions from './validatorsActions';
 import * as xhrActions from './xhrActions';
 import * as orderbookActions from './orderbookActions';
 import * as nftActions from './nftActions';
-import * as statsActions from './statsActions';
+import * as networkActions from './networkActions';
 
 export {
   accountsActions,
@@ -24,6 +24,6 @@ export {
   validatorsActions,
   xhrActions,
   orderbookActions,
+  networkActions,
   nftActions,
-  statsActions,
 };