+++ src/redux/hooks/useStats.js
@@ -1,6 +1,0 @@
-import useRedux from './useRedux';
-import { statsActions } from '../actions';
-
-const useStats = () => useRedux('statsReducer', statsActions);
-
-export default useStats;