+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Add IBC UI #130
 - Migrate orderbook to gecko API #255
 - Add 24 hr changes to validator list view #263
+- Add chain params to network dropdown #144
 
 ## 2.5.0